requirements.txt
langchain
langchain-groq
python-dotenv
langchain_community
langchain_core
streamlit
streamlit-webrtc
soundfile
matplotlib 
av 
numpy
openai-whisper